pyproject.toml
[project]
name = "reps"
version = "0.2.0"
description = "Resistance Exercise Pattern Sensing: Human activity recognition of resistance training exercise via accelerometer and gyroscope signals"
readme = "README.md"
authors = [
    { name = "MelaroJ", email = "melarojake@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "black>=25.1.0",
    "duckdb>=1.2.2",
    "mypy>=1.15.0",
    "polars>=1.27.1",
    "pydantic>=2.11.3",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pyyaml>=6.0.2",
    "ruff>=0.11.5",
]

[project.scripts]
reps = "reps:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
addopts = "-qx --cov=src --cov-report=xml --cov-report=term-missing"
testpaths = ["tests"]
